Improve copy label
